pyproject.toml
[tool.poetry]
name = "team9"
version = "0.1.0"
description = ""
authors = ["nyh <dkzndk1100@naver.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
django-dotenv = "^1.4.2"
django-restframework = "^0.0.1"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
dj-rest-auth = "^2.2.3"
PyMySQL = "^1.0.2"
django-sslserver = "^0.22"
cryptography = "^36.0.1"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
flake8 = "^4.0.1"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
